+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [Unreleased]
+- Fix passing arguments to etag_func and last_modified_func
 
 ## 0.1.0
 - Add first implementation